Sleepwatcher scripts.
This is an attempt to clear out dead SSH control sockets.
It's not perfect, but seems sensible. Unfortunately,
it didn't solve the problem - hanging ssh connections
the first time a connection is initiated.

Then again, those hangs only appear with GitHub. Perhaps
I just switch off the multiplexing with GitHub. But that's
where it's most useful...